Switch to H2 to reduce disk limits

+++ src/main/java/uk/co/tertiarybrewery/brewapi/pidintegration/TargetTempResponse.java
@@ -2,22 +2,22 @@ package uk.co.tertiarybrewery.brewapi.pidintegration;
 
 
 public class TargetTempResponse {
-    public float getActive() {
+    public double getActive() {
         return active;
     }
 
-    public void setActive(float active) {
+    public void setActive(double active) {
         this.active = active;
     }
 
-    public float getPre_warm() {
+    public double getPre_warm() {
         return pre_warm;
     }
 
-    public void setPre_warm(float pre_warm) {
+    public void setPre_warm(double pre_warm) {
         this.pre_warm = pre_warm;
     }
 
-    float active;
-    float pre_warm;
+    double active;
+    double pre_warm;
 }
